+++ src/components/Chart/types.tsx
@@ -1,0 +1,8 @@
+export interface TradingData {
+  date: string;
+  open: string;
+  high: string;
+  low: string;
+  close: string;
+  volume: string;
+}
